IDETECT-793: refactoring
